feat: set userId on API requests

+++ src/pages/api/journeys/[id]/index.ts
@@ -1,26 +1,42 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
 import createClient from '@/libs/supabase/api'
+import type { NextApiRequest, NextApiResponse } from 'next'
 
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
   const supabase = createClient(req, res)
-  const { id } = req.query
+  const { id, userId } = req.query
 
   if (req.method === 'DELETE') {
-    await supabase.from('journeys').delete().eq('id', id!)
-    res.status(204).json({
-      message: 'Journey deleted',
-    })
+    await supabase.from('journeys').delete().eq('id', id!).eq('userId', userId!)
+
+    res.status(204).end()
   } else if (req.method === 'GET') {
-    const { data: journey } = await supabase
+    const { data: journey, error: journeyError } = await supabase
       .from('journeys')
       .select('*')
       .eq('id', id!)
+      .eq('userId', userId!)
       .single()
 
-    res.status(200).json(journey)
+    if (!journey || journeyError) {
+      return res.status(404).json({
+        message: 'Journey not found',
+      })
+    }
+
+    const { data: expenses } = await supabase
+      .from('expenses')
+      .select('*')
+      .eq('journeyId', id!)
+
+    const { data: days } = await supabase
+      .from('days')
+      .select('*')
+      .eq('journeyId', id!)
+
+    res.status(200).json({ journey, expenses, days })
   } else {
     res.status(405).json({
       message: 'Method not allowed',
